browser-use-web-app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
playwright==1.40.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1

# LLM integrations
openai==1.3.8
google-generativeai==0.3.2
anthropic==0.7.8

# File processing - Use precompiled wheels for Python 3.13
PyPDF2==3.0.1
# Use openpyxl instead of pandas to avoid C++ compiler issues
openpyxl==3.1.2

# Optional: for more advanced features
python-dotenv==1.0.0
httpx==0.25.2